+++ src/core/authorization/index.ts
@@ -1,2 +1,1 @@
 export * from './graphql.guard';
-export * from './dto';
