+++ src/core/bootstrap/bootstrap.module.ts
@@ -6,9 +6,9 @@ import { ProfileModule } from '@domain/community/profile/profile.module';
 import { TagsetModule } from '@domain/common/tagset/tagset.module';
 import { UserModule } from '@domain/community/user/user.module';
 import { BootstrapService } from './bootstrap.service';
-import { AuthorizationModule } from '@core/authorization/authorization.module';
 import { OrganizationModule } from '@domain/community/organization/organization.module';
 import { AgentModule } from '@domain/agent/agent/agent.module';
+import { AdminAuthorizationModule } from '@services/admin/authorization/admin.authorization.module';
 
 @Module({
   imports: [
@@ -17,7 +17,7 @@ import { AgentModule } from '@domain/agent/agent/agent.module';
     ProfileModule,
     TagsetModule,
     UserModule,
-    AuthorizationModule,
+    AdminAuthorizationModule,
     OrganizationModule,
     TypeOrmModule.forFeature([Ecoverse]),
   ],
