+++ src/domain/common/entity/authorizable-entity/authorizable.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { AuthorizationEngineModule } from '@services/platform/authorization-engine/authorization-engine.module';
+import { AuthorizableService } from './authorizable.service';
+
+@Module({
+  imports: [AuthorizationEngineModule],
+  providers: [AuthorizableService],
+  exports: [AuthorizableService],
+})
+export class AuthorizableEntityModule {}
