+++ src/services/admin/authorization/admin.authorization.resolver.queries.ts
@@ -1,10 +1,6 @@
 import { CurrentUser, Profiling } from '@src/common/decorators';
 import { Args, Query, Resolver } from '@nestjs/graphql';
-import { AuthorizationService } from './authorization.service';
-import {
-  GraphqlGuard,
-  UsersWithAuthorizationCredentialInput,
-} from '@core/authorization';
+import { GraphqlGuard } from '@core/authorization';
 import { IUser } from '@domain/community/user';
 import { AuthorizationPrivilege, AuthorizationRoleGlobal } from '@common/enums';
 import { UserAuthorizationPrivilegesInput } from './dto/authorization.dto.user.authorization.privileges';
@@ -12,17 +8,21 @@ import { UseGuards } from '@nestjs/common/decorators/core/use-guards.decorator';
 import { IAuthorizationPolicy } from '@domain/common/authorization-policy/authorization.policy.interface';
 import { AuthorizationEngineService } from '@services/platform/authorization-engine/authorization-engine.service';
 import { AgentInfo } from '@core/authentication/agent-info';
+import { AdminAuthorizationService } from './admin.authorization.service';
+import { UsersWithAuthorizationCredentialInput } from './dto/authorization.dto.users.with.credential';
+import { AuthorizationPolicyService } from '@domain/common/authorization-policy/authorization.policy.service';
 
 @Resolver()
-export class AuthorizationResolverQueries {
+export class AdminAuthorizationResolverQueries {
   private authorizationQueriesPolicy: IAuthorizationPolicy;
 
   constructor(
+    private authorizationPolicyService: AuthorizationPolicyService,
     private authorizationEngine: AuthorizationEngineService,
-    private authorizationService: AuthorizationService
+    private adminAuthorizationService: AdminAuthorizationService
   ) {
     this.authorizationQueriesPolicy =
-      this.authorizationEngine.createGlobalRolesAuthorizationPolicy(
+      this.authorizationPolicyService.createGlobalRolesAuthorizationPolicy(
         [AuthorizationRoleGlobal.REGISTERED],
         [AuthorizationPrivilege.READ]
       );
@@ -46,7 +46,7 @@ export class AuthorizationResolverQueries {
       AuthorizationPrivilege.READ,
       `authorization query: ${agentInfo.email}`
     );
-    return await this.authorizationService.usersWithCredentials(
+    return await this.adminAuthorizationService.usersWithCredentials(
       credentialsCriteriaData
     );
   }
@@ -69,7 +69,7 @@ export class AuthorizationResolverQueries {
       AuthorizationPrivilege.READ,
       `authorization query: ${agentInfo.email}`
     );
-    return await this.authorizationService.userAuthorizationPrivileges(
+    return await this.adminAuthorizationService.userAuthorizationPrivileges(
       userAuthorizationPrivilegesData
     );
   }
