+++ src/services/admin/authorization/dto/index.ts
@@ -1,4 +1,3 @@
 export * from './authorization.dto.credential.grant';
 export * from './authorization.dto.credential.revoke';
 export * from './authorization.dto.users.with.credential';
-export * from '../../../domain/challenge/challenge/dto/challenge.dto.authorize.state.modification';
